possible fix for a truncated NSTask output issue
once in a few times, the output from a git command
(fetch, log, config or whatever) was being
returned truncated, resulting e.g. in missing
commits or a git path error.

it looks like this fixes the problem - if the data
is read synchronously instead of waiting for
notifications to come, nothing seems to be lost.